Python/requirements.txt
Flask==2.1.1
flask-cors==3.0.10
pytesseract==0.3.10
Pillow==8.4.0
PyMuPDF==1.19.0
concurrent==1.0.0
